requirements.txt
fastapi
uvicorn
celery
redis
psycopg2-binary
slack_sdk
google-generativeai
pymupdf
requests